Clarify origin of MnemonicManager
